design: 메인 - 사이드바 반응형 디자인 시안 맞춤 및 푸터 컴포넌트로 분리
