+++ src/components/Layout/FooterSection/FooterSection.module.scss
@@ -1,0 +1,103 @@
+@use "@/styles/globals.scss" as *;
+
+.footer {
+  display: flex;
+  flex-direction: column;
+  gap: 32px;
+  padding: 16px 0;
+  position: relative;
+
+  @include Size("mobile") {
+    padding: 0;
+    width: 100%;
+    gap: 30px;
+  }
+
+  .footerItems {
+    display: flex;
+    flex-direction: column;
+    gap: 12px;
+  }
+
+  .dropdown {
+    position: absolute;
+    padding: 8px 0;
+    bottom: 120px;
+    left: 16px;
+    background-color: $gray0;
+    box-shadow: 0 0px 12px rgba(0, 0, 0, 0.15);
+    border-radius: 12px;
+    overflow: hidden;
+    min-width: 180px;
+    z-index: 10;
+
+    @include Size("tablet") {
+      bottom: 80px;
+      left: 0px;
+    }
+
+    @include Size("mobile") {
+      bottom: auto;
+      top: -80px;
+      left: auto;
+    }
+
+    .dropdownItem {
+      @include Sub4;
+      display: block;
+      width: 100%;
+      padding: 12px 16px;
+      text-align: left;
+      background: none;
+      border: none;
+      cursor: pointer;
+      font-size: 14px;
+      color: $gray80;
+      text-decoration: none;
+      transition: background-color 0.2s;
+
+      &:hover {
+        background-color: $gray20;
+      }
+    }
+  }
+
+  .subLinkWrapper {
+    display: flex;
+    flex-direction: column;
+    gap: 4px;
+    padding-left: 16px;
+    @include Label3;
+    color: $gray60;
+
+    @include Size("mobile") {
+      padding-left: 0;
+      @include Cap2;
+    }
+
+    .subLink {
+      display: flex;
+      align-items: center;
+
+      a {
+        color: $gray60;
+        text-decoration: none;
+        transition: color 0.2s;
+
+        &:hover {
+          color: $gray80;
+        }
+
+        &:not(:last-child)::after {
+          content: "•";
+          margin: 0 6px;
+          color: $gray50;
+        }
+      }
+    }
+
+    @include Size("tablet") {
+      display: none;
+    }
+  }
+}
