+++ src/components/Layout/Header/Header.module.scss
@@ -403,81 +403,4 @@
   flex-direction: column-reverse;
   align-items: flex-start;
   margin-top: 24px;
-
-  .footerMenu {
-    display: flex;
-    flex-direction: column;
-    gap: 12px;
-    padding-bottom: 24px;
-  }
-
-  .dropdown {
-    position: absolute;
-    cursor: pointer;
-    background: $gray0;
-    box-shadow: 0px 2px 8px rgba(0, 0, 0, 0.08);
-    border-radius: 8px;
-    z-index: 90px;
-    top: -80px;
-    right: 0;
-    width: 200px;
-
-    .dropdownItem {
-      display: flex;
-      align-items: flex-start;
-      cursor: pointer;
-      @include Sub4;
-      color: $gray80;
-      text-decoration: none;
-      width: 100%;
-      outline: none;
-      border: none;
-
-      &:hover {
-        background-color: $gray20;
-      }
-    }
-  }
-
-  .infoWrapper {
-    display: flex;
-    flex-direction: column;
-    gap: 4px;
-
-    .item {
-      display: flex;
-      align-items: center;
-      gap: 6px;
-      height: 17px;
-      padding: 0;
-    }
-
-    .subLink {
-      display: flex;
-      gap: 6px;
-      color: $gray60;
-      @include Cap2;
-
-      a {
-        color: $gray60;
-        text-decoration: none;
-
-        &:hover {
-          text-decoration: underline;
-        }
-      }
-
-      a:not(:last-child)::after {
-        content: "  \00B7";
-        margin: 0 2px;
-        color: $gray50;
-      }
-    }
-
-    .copy {
-      width: fit-content;
-      color: $gray60;
-      @include Cap2;
-    }
-  }
 }
