+++ src/components/Layout/Header/Header.tsx
@@ -8,15 +8,13 @@ import { useModalStore } from "@/states/modalStore";
 import { useAuthStore } from "@/states/authStore";
 import { useMyData } from "@/api/users/getMe";
 import { useRouter } from "next/router";
-import { useToast } from "@/hooks/useToast";
 import Notifications from "@/components/Notifications/Notifications";
 import { useDeviceStore } from "@/states/deviceStore";
 import { useIsMobile } from "@/hooks/useIsMobile";
 import useAuthCheck from "@/hooks/useAuthCheck";
 import { usePreventScroll } from "@/hooks/usePreventScroll";
 import axiosInstance from "@/constants/baseurl";
-import { FOOTER_ITEMS } from "@/constants/footer";
-import SidebarFooterItem, { FooterIconName } from "../Sidebar/SidebarFooterItem/SidebarFooterItem";
+import FooterSection from "@/components/Layout/FooterSection/FooterSection";
 
 export default function Header() {
   const openModal = useModalStore((state) => state.openModal);
@@ -30,16 +28,13 @@ export default function Header() {
   const indicatorRef = useRef<HTMLDivElement>(null);
   const [isDropdownOpen, setIsDropdownOpen] = useState(false);
   const dropdownRef = useRef<HTMLDivElement>(null);
-  const askDropdownRef = useRef<HTMLDivElement>(null);
   const notificationRef = useRef<HTMLDivElement>(null);
   const [showNotifications, setShowNotifications] = useState(false);
-  const { showToast } = useToast();
   const router = useRouter();
   const isMobile = useDeviceStore((state) => state.isMobile);
   const isTablet = useDeviceStore((state) => state.isTablet);
   const [isMenuOpen, setIsMenuOpen] = useState(false);
   const [isSubMenuOpen, setIsSubMenuOpen] = useState(false);
-  const [isAskDropdownOpen, setIsAskDropdownOpen] = useState(false);
   const isPostPage = ["/board", "/board/write", "/posts/[id]", "/posts/[id]/edit"].includes(
     router.pathname,
   );
@@ -52,7 +47,6 @@ export default function Header() {
   const [isHome, setIsHome] = useState(false);
   useAuthCheck(setIsHome);
   useIsMobile();
-  const email = "grimity.official@gmail.com";
   const navItems = [
     { name: "홈", path: "/" },
     { name: "인기그림", path: "/popular" },
@@ -86,16 +80,6 @@ export default function Header() {
       };
     }
   }, [isMobile, showNotifications]);
-
-  const copyToClipboard = async () => {
-    try {
-      await navigator.clipboard.writeText(email);
-      showToast("이메일이 복사되었습니다!", "success");
-    } catch (error) {
-      console.error("클립보드 복사 실패:", error);
-      showToast("복사에 실패했습니다.", "success");
-    }
-  };
 
   const handleNavClick = (item: { name: string; path: string }) => {
     setActiveNav(item.name);
@@ -164,19 +148,6 @@ export default function Header() {
     }
   }, [activeNav]);
 
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (askDropdownRef.current && !askDropdownRef.current.contains(event.target as Node)) {
-        setIsAskDropdownOpen(false);
-      }
-    };
-
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, []);
-
   const handleClickLogo = () => {
     if (router.pathname === "/") {
       router.reload();
@@ -200,15 +171,6 @@ export default function Header() {
 
   const toggleSubmenu = () => {
     setIsSubMenuOpen((prev) => !prev);
-  };
-
-  const handleFooterClick = (itemIcon: FooterIconName, route?: string) => {
-    if (route) {
-      closeMenu();
-      router.push(route);
-    } else if (itemIcon === "ask") {
-      setIsAskDropdownOpen(!isAskDropdownOpen);
-    }
   };
 
   usePreventScroll(isMenuOpen || (isMobile && showNotifications));
@@ -508,65 +470,7 @@ export default function Header() {
                   </div>
 
                   <div className={styles.mobileFooterWrap}>
-                    <div className={styles.footerMenu}>
-                      {FOOTER_ITEMS.map((item, index) => (
-                        <SidebarFooterItem
-                          key={index}
-                          icon={item.icon as FooterIconName}
-                          label={item.label}
-                          onClickItem={() => handleFooterClick(item.icon, item.route)}
-                          isHaveDropdown={item.isHaveDropdown}
-                          isDropdownOpen={item.icon === "ask" ? isAskDropdownOpen : false}
-                        />
-                      ))}
-                    </div>
-                    {isAskDropdownOpen && (
-                      <div className={styles.dropdown} ref={askDropdownRef}>
-                        <a
-                          href="https://open.kakao.com/o/sKYFewgh"
-                          target="_blank"
-                          rel="noopener noreferrer"
-                          className={styles.dropdownItem}
-                          onClick={closeMenu}
-                        >
-                          카카오톡으로 문의하기
-                        </a>
-                        <div
-                          onClick={() => {
-                            copyToClipboard();
-                            closeMenu();
-                          }}
-                          className={styles.dropdownItem}
-                        >
-                          메일로 보내기
-                        </div>
-                      </div>
-                    )}
-                    <div className={styles.infoWrapper}>
-                      <div className={styles.item}>
-                        <div className={styles.subLink}>
-                          <a
-                            href="https://nostalgic-patch-498.notion.site/1930ac6bf29881b9aa19ff623c69b8e6?pvs=74"
-                            target="_blank"
-                            rel="noopener noreferrer"
-                            onClick={closeMenu}
-                          >
-                            이용약관
-                          </a>
-                          <a
-                            href="https://nostalgic-patch-498.notion.site/1930ac6bf29881e9a3e4c405e7f49f2b?pvs=73"
-                            target="_blank"
-                            rel="noopener noreferrer"
-                            onClick={closeMenu}
-                          >
-                            개인정보취급방침
-                          </a>
-                        </div>
-                      </div>
-                      <div className={styles.item}>
-                        <p className={styles.copy}>© Grimity. All rights reserved.</p>
-                      </div>
-                    </div>
+                    <FooterSection onClose={closeMenu} />
                   </div>
                 </div>
               </div>
