+++ src/components/Layout/Sidebar/Sidebar.module.scss
@@ -10,7 +10,7 @@
   top: 0;
   left: 0;
   margin-top: 60px;
-  padding: 0 10px;
+  padding: 0 12px;
   height: 100vh;
   background-color: $gray0;
   border-right: 1px solid $gray30;
@@ -33,93 +33,10 @@
     margin-bottom: 80px;
     width: 100%;
 
-    .subLink {
-      display: block;
-      flex-direction: column;
-      align-items: flex-start;
-      width: 100%;
-      color: $gray60;
-      @include Label3;
-      margin-top: 24px;
-      padding-left: 20px;
-
-      a {
-        list-style: none;
-        text-decoration: none;
-        color: $gray60;
-
-        &::after {
-          color: $gray50;
-          content: " • ";
-          padding-left: 6px;
-        }
-        &:last-child::after {
-          content: none;
-        }
-      }
-
-      @include Size("tablet") {
-        display: none;
-      }
-    }
-
-    .copy {
-      @include Label3;
-      color: $gray60;
-      margin-top: 12px;
-      padding-left: 20px;
-
-      @include Size("tablet") {
-        display: none;
-      }
-    }
-
     @include Size("tablet") {
       max-width: 80px;
       justify-content: flex-start;
       align-items: center;
     }
-  }
-}
-
-.footerItems {
-  display: flex;
-  flex-direction: column;
-  row-gap: 12px;
-}
-
-.dropdown {
-  position: absolute;
-  padding: 8px 0;
-  bottom: 190px;
-  left: 36px;
-  background: $gray0;
-  box-shadow: 0px 0px 12px rgba(0, 0, 0, 0.15);
-  border-radius: 12px;
-  z-index: 10;
-  width: 180px;
-  overflow: hidden;
-
-  @include Size("tablet") {
-    bottom: 160px;
-    left: 16px;
-  }
-}
-
-.dropdownItem,
-button.dropdownItem {
-  all: unset;
-  display: flex;
-  align-items: center;
-  padding: 12px 16px;
-  cursor: pointer;
-  @include Sub4;
-  color: $gray80;
-  text-decoration: none;
-  width: 100%;
-  text-align: left;
-
-  &:hover {
-    background-color: $gray20;
   }
 }
