+++ src/components/Layout/Sidebar/Sidebar.tsx
@@ -1,20 +1,18 @@
-import { useState, useRef, useEffect } from "react";
 import { useRouter } from "next/router";
-import SidebarItem, { BaseIconName } from "./SidebarItem/SidebarItem";
-import SidebarFooterItem, { FooterIconName } from "./SidebarFooterItem/SidebarFooterItem";
-import styles from "./Sidebar.module.scss";
-import { MENU_ITEMS } from "@/constants/menu";
-import { FOOTER_ITEMS } from "@/constants/footer";
+
 import { useAuthStore } from "@/states/authStore";
 import { useDeviceStore } from "@/states/deviceStore";
+
+import SidebarItem, { BaseIconName } from "@/components/Layout/Sidebar/SidebarItem/SidebarItem";
+import FooterSection from "@/components/Layout/FooterSection/FooterSection";
+
 import { useIsMobile } from "@/hooks/useIsMobile";
-import { useToast } from "@/hooks/useToast";
+
+import { MENU_ITEMS } from "@/constants/menu";
+
+import styles from "@/components/Layout/Sidebar/Sidebar.module.scss";
 
 const Sidebar = () => {
-  const [isAskDropdownOpen, setIsAskDropdownOpen] = useState(false);
-  const dropdownRef = useRef<HTMLDivElement>(null);
-  const { showToast } = useToast();
-  const email = "grimity.official@gmail.com";
   const { isLoggedIn } = useAuthStore((state) => state);
   const router = useRouter();
   const { isMobile } = useDeviceStore((state) => state);
@@ -24,28 +22,8 @@ const Sidebar = () => {
     ? [...MENU_ITEMS, { icon: "following", label: "팔로잉", route: "following" }]
     : MENU_ITEMS;
 
-  const footerItems = FOOTER_ITEMS;
-
   const handleItemClick = (route: string) => {
     router.push(`/${route}`);
-  };
-
-  const handleFooterClick = (itemIcon: FooterIconName, route?: string) => {
-    if (route) {
-      router.push(route);
-    } else if (itemIcon === "ask") {
-      setIsAskDropdownOpen(!isAskDropdownOpen);
-    }
-  };
-
-  const copyToClipboard = async () => {
-    try {
-      await navigator.clipboard.writeText(email);
-      showToast("이메일이 복사되었습니다!", "success");
-    } catch (error) {
-      console.error("클립보드 복사 실패:", error);
-      showToast("복사에 실패했습니다.", "success");
-    }
   };
 
   const isItemActive = (route: string) => {
@@ -59,19 +37,6 @@ const Sidebar = () => {
 
     return false;
   };
-
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (dropdownRef.current && !dropdownRef.current.contains(event.target as Node)) {
-        setIsAskDropdownOpen(false);
-      }
-    };
-
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, []);
 
   return (
     <>
@@ -90,50 +55,7 @@ const Sidebar = () => {
           </div>
 
           <div className={styles.footer}>
-            <div className={styles.footerItems}>
-              {footerItems.map((item, index) => (
-                <SidebarFooterItem
-                  key={index}
-                  icon={item.icon as FooterIconName}
-                  label={item.label}
-                  onClickItem={() => handleFooterClick(item.icon, item.route)}
-                  isHaveDropdown={item.isHaveDropdown}
-                  isDropdownOpen={isAskDropdownOpen}
-                />
-              ))}
-            </div>
-            {isAskDropdownOpen && (
-              <div className={styles.dropdown} ref={dropdownRef}>
-                <a
-                  href="https://open.kakao.com/o/sKYFewgh"
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  className={styles.dropdownItem}
-                >
-                  카카오톡으로 문의하기
-                </a>
-                <button onClick={copyToClipboard} className={styles.dropdownItem}>
-                  메일로 보내기
-                </button>
-              </div>
-            )}
-            <div className={styles.subLink}>
-              <a
-                href="https://nostalgic-patch-498.notion.site/1930ac6bf29881b9aa19ff623c69b8e6?pvs=74"
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                개인정보취급방침
-              </a>
-              <a
-                href="https://nostalgic-patch-498.notion.site/1930ac6bf29881e9a3e4c405e7f49f2b?pvs=73"
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                이용약관
-              </a>
-            </div>
-            <p className={styles.copy}>© Grimity. All rights reserved.</p>
+            <FooterSection />
           </div>
         </nav>
       )}
