+++ src/components/Layout/Sidebar/SidebarFooterItem/SidebarFooterItem.module.scss
@@ -5,7 +5,7 @@
   align-items: center;
   justify-content: flex-start;
   max-width: 234px;
-  padding: 0 20px;
+  padding: 0 16px;
   @include Body1;
   color: $gray60;
   cursor: pointer;
