+++ src/components/Layout/Sidebar/SidebarItem/SidebarItem.module.scss
@@ -6,7 +6,7 @@
   justify-content: flex-start;
   width: 234px;
   height: 50px;
-  padding: 0 20px;
+  padding: 0 16px;
   color: $gray60;
   cursor: pointer;
   transition: background 0.3s, color 0.3s;
